+++ JiraAssistant/Model/Jira/PivotJiraIssue.cs
@@ -26,7 +26,7 @@ namespace JiraAssistant.Model.Jira
       public DateTime Resolved { get; private set; }
       public string Assignee { get; private set; }
       public string Reporter { get; private set; }
-      public int StoryPoints { get; private set; }
+      public float StoryPoints { get; private set; }
       public string Priority { get; private set; }
       public string Type { get; private set; }
       public string Resolution { get; private set; }
